move Input Component to renderInput
