+++ src/app/renders/renderInput.js
@@ -1,5 +1,5 @@
 import React from "react";
-import {Form, Label, Transition} from "semantic-ui-react";
+import {Form, Label, Transition, Checkbox} from "semantic-ui-react";
 
 const renderInput = field => (
     <>
@@ -37,4 +37,55 @@ const renderInput = field => (
     </>
 );
 
-export default renderInput;
+const renderCheckbox = ({
+    input: { value, onChange, ...input },
+    meta: { touched, error },
+    ...rest
+  }) => (
+    <Form.Field error={!!(touched && error)}>
+      <Checkbox
+        {...input}
+        {...rest}
+        defaultChecked={!!value}
+        onChange={(e, data) => onChange(data.checked)}
+        type="checkbox"        
+      />
+      {/* {touched && error && <span>{error}</span>} */}
+    </Form.Field>
+  );
+
+  const radioGroup = ({input: {onChange, name, value}, meta: {touched, error}, options}) => {
+
+      const radioSet = options.map(opt => (
+          <Form.Radio 
+                checked={value === opt.radioValue}
+                label={opt.label}
+                name={name}
+                onChange={() => onChange(opt.radioValue)}
+          />
+      ))
+      return (
+        <Form.Group inline>
+        <label>Genders</label>
+        <Form.Field error={!!(touched && error)}>
+            {radioSet}
+          </Form.Field>
+        </Form.Group>         
+      )
+  }
+
+const renderRadio = field => (
+    <Form.Radio
+        checked={field.input.value === field.radioValue}
+        label={field.label}
+        name={field.input.name}
+        onChange={() => field.input.onChange(field.radioValue)}
+    />
+);
+
+export default {
+    renderInput, 
+    radioGroup,
+    renderCheckbox
+    
+ };
